Combine `IframeOutput` into `MaybeCellOutputs`
